+++ Models/Employee.cs
@@ -4,13 +4,13 @@ namespace SalaryCalculator.Models
     public class Employee
     {
         public int Id { get; }
-        public double BasicSalary { get; }
-        public Employee(int id, double basicSalary)
+        public decimal BasicSalary { get; }
+        public Employee(int id, decimal basicSalary)
         {
             this.Id = id;
             this.BasicSalary = basicSalary;
         }
-        public virtual double GetTotalSalary()
+        public virtual decimal GetTotalSalary()
         {
             return BasicSalary;
         }
